feat(mutant description): show optional mutant description (#37)
Add an optional `description` field to a mutant in the schema. Coincidently, make the `replacement` field optional. This effectively adds better support for PITtest (where the replacement is missing, instead they show a description).

When the `replacement` is missing from a mutant, the mutator name is shown instead (see "📖 Show more" example below).

If there is a description, it is shown in the popup as a "📖 Show more" link:

![show more link example](https://user-images.githubusercontent.com/5868064/57451579-fe93bd00-7261-11e9-9219-f12f688f3efc.png)

If you click on it, you'll see the additional description in a modal dialog:

![modal dialog example](https://user-images.githubusercontent.com/5868064/57451711-55999200-7262-11e9-8c40-201d3eb7d306.png)