feat(n/a mutation score): support no mutation score
Allow a mutation score *N/A* when there is no score to be calculated for a file or a directory. It is shown without a mutation score:

![image](https://user-images.githubusercontent.com/10114577/74157960-822b7180-4c19-11ea-84a3-507de922ae10.png)

Both in Scala as in JavaScript this is implemented as mutation score `NaN`. 

Breakingish change: mutation score is now `NaN` instead of `0` when no mutation score could be calculated.