feat(metrics): add test metrics
Add metrics about tests to `mutation-testing-metrics`. Now exporting a new function called `calculateMutationTestMetrics`. It requires a `MutationTestResult` as input and and calculates both the metrics for files under test as well as test files.

Currently sporting 2 metrics for a test file, namely the `total` tests and the number of "pacifist" tests, meaning tests that didn't kill any mutants.