feat(schema): add test file source and status reason
Add `statusReason` to a mutant result and `source` to a test file.

Status reason:
The reason that this mutant has this status. In case of a killed mutant, this should be filled with the failure message(s) of the failing tests. In case of a error mutant, this should be filled with the error message.

Source:
Full source code of the test file, this can be used to display in the report.