feat(schema): add test file source and status reason (#893)
Add `statusReason` to a mutant result and `source` to a test file.

Status reason:
The reason that this mutant has this status. In the case of a killed mutant, this should be filled with the failure message(s) of the failing tests. In case of an error mutant, this should be filled with the error message.

Closes #891 

Source:
Full source code of the test file. This can be used to display in the report.

Closes #892