feat(test-view): add test view with test details (#1014)
Add a "🧪 Test" view to the mutation test report. In this view, you can find information about tests.

It functions the same way as the already existing view, which is now named the "👽 Mutant" view.

When there is information about tests in the report JSON, `mutation-testing-elements` will now render 2 nav tabs (top left). Here you can switch between the 2 views. The mutant view will open by default.

![image](https://user-images.githubusercontent.com/1828233/115002945-94501200-9ea5-11eb-8bf6-f63d3e2aeb15.png)

When opening the test view, you either get the file/directory view or a big list of all tests, depending on the data available. 

* If the report JSON contains tests without a test file (i.e. `{ "testFiles": { "": { "tests": [/*...*/] } }`), it will show a big list of tests
* If the report JSON contains the test file names (i.e. `{ "testFiles": { "foo.spec.js": { "tests": [/*...*/] } }`), it will show the directory view

When navigating to a file view, either the source of that file is displayed with the tests in them, or a list of tests is displayed when there is no source available.