perf(elements): improve performance of real-time reporting (#2498)
Instead of recalculating every time an event comes through, only react to event updates at an interval. By using the `debounceTime` function from rxjs we can tell rxjs to ignore events for a given time period, throttling the amount of times we have to recalculate and redraw the report elements.